Changed the type of balance to float

+++ app/model/User.js
@@ -4,7 +4,7 @@ Ext.define('RiskiMarket.model.User', {
     fields: [
         { name: 'key',     type: 'string' },
         { name: 'name',    type: 'string' },
-        { name: 'balance', type: 'int'    },
+        { name: 'balance', type: 'float'  },
         { name: 'admin',   type: 'bool'   }
     ]
 });
